Split DateOnlyAssertionSpecs
